apps/contacts/serailizer.py
from rest_framework import serializers

from apps.contacts.models import ContactUs


class ContactUsSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactUs
        fields = '__all__'
